requirements.txt
python-chess
openai
outlines
transformers
